Added a NamingStrategy class to Entity provider
FireBird works better (IMO) with table and column names in upper case
and without quotes.
The default behavior with EF if to use pascal case and double quotes but
some times this can be a problem.
So I added a NamingStrategy class that handles the QuoteIdentifier()
method.
In future updates the naming decoration should be independent for table
and column names.