Update DbValue.cs
Add support for Ticks to the GetTime() method.
Some Orm frameworks like ServiceStack.OrmLite convert TimeOnly property  values to Ticks and pass a Int64 value when creating parameters, instead of passing a TimeOnly value.
With this change it will be possible to use this kind of frameworks without getting conversion errors.